A possible approach to generating autoloads.
This is essentially a variation on the approach suggested by @cpitclaudel.
The idea is that, since `package-install*' will call `update-directory-autoloads'
in the top-level directory of the package being installed exactly once
at installation time, code to "bootstrap" autoloads in sub-directories can be put into
a *.el file in the top directory of the package.